Fix: #1032 Renamed ButtonGroup to ButtonSegmented (#1072)
BREAKING CHANGE: The Component ButtonGroup is renamed as ButtonSegmented to align with fundamental-styles
