+++ src/Button/__stories__/ButtonSegmented.stories.js
@@ -1,26 +1,26 @@
 /* eslint-disable react/no-multi-comp */
 import Button from '../Button';
-import ButtonGroup from '../ButtonGroup';
+import ButtonSegmented from '../ButtonSegmented';
 import React from 'react';
 
 export default {
-    title: 'Component API/ButtonGroup',
-    component: ButtonGroup
+    title: 'Component API/ButtonSegmented',
+    component: ButtonSegmented
 };
 
-export const primary = () => (<ButtonGroup>
+export const primary = () => (<ButtonSegmented>
     <Button glyph='survey' />
     <Button glyph='pie-chart' selected />
     <Button glyph='pool' />
-</ButtonGroup>);
+</ButtonSegmented>);
 
 
 export const compact = () => (
-    <ButtonGroup>
+    <ButtonSegmented>
         <Button compact>Left</Button>
         <Button compact selected>Middle</Button>
         <Button compact>Right</Button>
-    </ButtonGroup>
+    </ButtonSegmented>
 );
 
 compact.story = {
